# backend/docs/swagger.yaml
definitions:
  main.CreateTaskPayload:
    properties:
      deadline:
        type: string
      description:
        type: string
      priority:
        enum:
        - Alta
        - Média
        - Baixa
        type: string
      status:
        enum:
        - A Fazer
        - Em Progresso
        - Concluídas
        type: string
      title:
        minLength: 3
        type: string
    required:
    - title
    type: object
  main.Task:
    properties:
      createdAt:
        type: string
      deadline:
        type: string
      description:
        description: If no description, it can be ignored by JSON
        type: string
      id:
        type: string
      priority:
        enum:
        - Alta
        - Média
        - Baixa
        type: string
      status:
        enum:
        - A Fazer
        - Em Progresso
        - Concluídas
        type: string
      title:
        description: The title is mandatory and must have at least 3 characters.
        minLength: 3
        type: string
      updatedAt:
        type: string
    required:
    - title
    type: object
  main.UpdateTaskPayload:
    properties:
      deadline:
        type: string
      description:
        type: string
      priority:
        enum:
        - Alta
        - Média
        - Baixa
        type: string
      status:
        enum:
        - A Fazer
        - Em Progresso
        - Concluídas
        type: string
      title:
        minLength: 3
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para o desafio fullstack da Veritas (React + Go).
  termsOfService: 'http:'
  title: API Kanban de Tarefas (Veritas)
  version: "1.0"
paths:
  /tasks:
    get:
      description: Retorna em uma lista todas as Tasks encontradas
      parameters:
      - description: Busca por parte do titulo fornecida(case-insensitive)
        in: query
        name: task
        type: string
      - description: Ordena por prioridade fornecida(priority_grw, priority_dec)
        in: query
        name: ordem
        type: string
      - description: Filtra por prioridade (Alta, Media ou Baixa)
        in: query
        name: filtro
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Task'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Lista as Tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Preenche os dados da task e cria ela
      parameters:
      - description: Payload da Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/main.CreateTaskPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cria uma nova task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Remove uma Task relacionada ao unico ID
      parameters:
      - description: ID da Task
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deleta uma Task
      tags:
      - tasks
    get:
      description: Retorna a task relacionada ao unico ID
      parameters:
      - description: ID da Task
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Busca uma task por um ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Atualiza os campos de uma Task (titulo, descricao, prioridade,
        prazo, status)
      parameters:
      - description: ID da Task
        in: path
        name: id
        required: true
        type: string
      - description: Campos da Task para Atualizar
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/main.UpdateTaskPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualiza uma Task existente
      tags:
      - tasks
swagger: "2.0"